Add rt-kernel bsp-specific sample-app file
Split rt-kernel sample-app into a board-specific file and a generic
file, so that p-net can be built for any rt-kernel board.
